add registration through mail

+++ BusinessService/src/test/java/com/company/innowise/airticketsapp/businessservice/service/FlightServiceTest.java
@@ -4,7 +4,6 @@ import com.company.innowise.airticketsapp.businessservice.dto.NewFlightDto;
 import com.company.innowise.airticketsapp.businessservice.dto.UpdatedFlightDto;
 import com.company.innowise.airticketsapp.businessservice.model.Airport;
 import com.company.innowise.airticketsapp.businessservice.model.Flight;
-import com.company.innowise.airticketsapp.businessservice.model.Ticket;
 import com.company.innowise.airticketsapp.businessservice.repository.FlightRepository;
 import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentMatchers;
@@ -12,12 +11,9 @@ import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
-
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Optional;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 @SpringBootTest
 class FlightServiceTest {
@@ -51,7 +47,7 @@ class FlightServiceTest {
         newFlightDto.setToId(0);
         newFlightDto.setTimeDeparture(LocalDateTime.now());
         newFlightDto.setTimeArrive(LocalDateTime.now());
-        newFlightDto.setPrice(0d);
+        newFlightDto.setPrice(BigDecimal.valueOf(0d));
         flightService.addFlight(newFlightDto);
         Mockito.verify(flightRepository, Mockito.times(1))
                 .save(ArgumentMatchers.any(Flight.class));
@@ -76,4 +72,5 @@ class FlightServiceTest {
         Mockito.verify(flightRepository, Mockito.times(1))
                 .save(ArgumentMatchers.any(Flight.class));
     }
+
 }
